refactor: shifted socket package to server subdirectory
